+++ src/models/scriptModel.ts
@@ -6,12 +6,31 @@ export class Script {
     readonly relativePath: string;
     readonly name: string;
     readonly args: string[];
-    constructor(scrpitPath: string | vscode.Uri, args?: string[]) {
+
+    readonly isLocal: boolean;
+
+    constructor(scrpitPath: string | vscode.Uri, isLocal: boolean, args?: string[], ) {
         this.uri = scrpitPath instanceof vscode.Uri ? scrpitPath : vscode.Uri.file(scrpitPath);
-        this.relativePath = path.relative(vscode.workspace.workspaceFolders?.[0].uri.path ?? '~/', this.uri.path);
+
+        const workspacePath = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
+        const filePath = this.uri.fsPath;
+        let relativePath: string;
+        if (workspacePath) {
+            relativePath = path.relative(workspacePath, filePath);
+            const upLevelCount = (relativePath.match(/\.\.\//g) || []).length;
+            if (upLevelCount > 2) {
+                relativePath = filePath;
+            }
+        } else {
+            relativePath = filePath;
+        }
+        this.relativePath = relativePath;
+
         let uriPart = this.uri.path.split('/');
         this.name = uriPart[uriPart.length - 1];
+
         this.args = args ?? [];
+        this.isLocal = isLocal;
     }
 
     toString() {
@@ -24,6 +43,6 @@ export class Script {
     }
 
     public static fromObject(obj: Script): Script {
-        return new Script(obj.uri.path, obj.args);
+        return new Script(obj.uri.path, obj.isLocal, obj.args);
     }
 }