+++ src/services/configService.ts
@@ -51,6 +51,9 @@ class ConfigService {
     public get terminalShellList(): Array<string> {
         return vscode.workspace.getConfiguration('slurm--.launch.terminal').get('shellList') ?? ['bash', 'zsh', 'fish'];
     }
+    public get scriptsExtList(): Array<string> {
+        return vscode.workspace.getConfiguration('slurm--.launch.scripts').get('extensionList') ?? ['sh', 'slurm', 'sbatch'];
+    }
 
     #taskSortKey!: string;
     public get taskSortKey(): TaskSortKeys {